Update index.js and add export default to Arrow and Dots
